Add Beach Segment is mostly written. Need to figure out how to add edges to the tree.
Handling circle events is next.
Then finally, generating the graph from the half edges.

+++ VoronoiLib/Structures/FortuneEvent.cs
@@ -8,7 +8,7 @@ namespace VoronoiLib.Structures
 {
     interface FortuneEvent : IComparable<FortuneEvent>
     {
-        int X { get; }
-        int Y { get; }
+        double X { get; }
+        double Y { get; }
     }
 }
