+++ VoronoiLib/Structures/FortuneSite.cs
@@ -2,10 +2,11 @@
 {
     public class FortuneSite
     {
-        public int X { get; }
-        public int Y { get; }
+        public double X { get; }
+        public double Y { get; }
+        
 
-        public FortuneSite(int x, int y)
+        public FortuneSite(double x, double y)
         {
             X = x;
             Y = y;
